Add README file and other examples of Nebula
