+++ cli/jobs/nebulaml/cifar10_deepspeed/job.yml
