+++ cli/jobs/nebulaml/cifar10_deepspeed/src/ds_config.json
